reorganizes files
