+++ modules/components/src/AdvancedSqonBuilder/SqonEntryContent/BooleanOp.js
@@ -1,24 +1,9 @@
 import React from 'react';
 import Component from 'react-component-component';
-import {
-  resolveSyntheticSqon,
-  isReference,
-  isBooleanOp,
-  isFieldOp,
-  isEmptySqon,
-  DisplayNameMapContext,
-  removeSqonPath,
-  setSqonAtPath,
-  getOperationAtPath,
-} from './utils';
-import FieldFilter from './FieldFilter';
-import ClickAwayListener from '../utils/ClickAwayListener.js';
-
-const PillRemoveButton = ({ onClick }) => (
-  <span className={`pillRemoveButton`} onClick={onClick}>
-    ✕
-  </span>
-);
+import { isReference, isBooleanOp, isFieldOp, isEmptySqon } from '../utils';
+import FieldOp from './FieldOp';
+import ClickAwayListener from '../../utils/ClickAwayListener.js';
+import { PillRemoveButton } from './common';
 
 const SqonReference = ({ refIndex, onRemoveClick = () => {} }) => (
   <span className={`sqonReference pill`}>
@@ -66,77 +51,11 @@ const LogicalOpSelector = ({ opName, onChange = newOpName => {} }) => {
   );
 };
 
-const FieldOp = ({
-  onSqonChange = fullSqon => {},
-  onContentRemove = () => {},
-  fullSyntheticSqon,
-  sqonPath = [],
-}) => {
-  const fieldOpObj = getOperationAtPath(sqonPath)(fullSyntheticSqon);
-  const {
-    op,
-    content: { field, value },
-  } = fieldOpObj;
-  const initialState = { isOpen: false };
-  const onClickAway = s => () => {
-    s.setState({ isOpen: false });
-  };
-  const toggleDropdown = s => () => s.setState({ isOpen: !s.state.isOpen });
-  const onRemoveClick = () => {
-    onContentRemove(fieldOpObj);
-  };
-  const onNewSqonSubmitted = s => newSqon => {
-    onSqonChange(newSqon);
-    toggleDropdown(s)();
-  };
-  return (
-    <Component initialState={initialState}>
-      {s => (
-        <DisplayNameMapContext.Consumer>
-          {(fieldDisplayNameMap = {}) => (
-            <span className={`fieldOp pill`}>
-              <span className={'opContainer'}>
-                <span className={`fieldName`}>
-                  {fieldDisplayNameMap[field] || field}{' '}
-                </span>
-                <span className={`opName`}>{op} </span>
-              </span>
-              <ClickAwayListener
-                className={'selectionContainer'}
-                handler={onClickAway(s)}
-              >
-                <span className={'valueDisplay'} onClick={toggleDropdown(s)}>
-                  {Array.isArray(value) ? value.join(', ') : value}{' '}
-                </span>
-                <span
-                  onClick={toggleDropdown(s)}
-                  className={`fa fa-chevron-${s.state.isOpen ? 'up' : 'down'}`}
-                />
-                {s.state.isOpen && (
-                  <div className={`fieldFilterContainer`}>
-                    <FieldFilter
-                      sqonPath={sqonPath}
-                      initialSqon={fullSyntheticSqon}
-                      onSubmit={onNewSqonSubmitted(s)}
-                      onCancel={toggleDropdown(s)}
-                    />
-                  </div>
-                )}
-              </ClickAwayListener>
-              <PillRemoveButton onClick={onRemoveClick} />
-            </span>
-          )}
-        </DisplayNameMapContext.Consumer>
-      )}
-    </Component>
-  );
-};
-
 /**
  * BooleanOp handles nested sqons through recursive rendering.
  * This will be useful for supporting brackets later.
  */
-const BooleanOp = ({
+export default ({
   contentPath = [],
   onFieldOpRemove = path => {},
   onChange = (changedPath, newOp) => {},
@@ -193,26 +112,5 @@ const BooleanOp = ({
         );
       })}
     </span>
-  );
-};
-
-export default ({ syntheticSqon, onSqonChange, allSyntheticSqons = [] }) => {
-  const onFieldOpRemove = removedPath =>
-    onSqonChange(removeSqonPath(removedPath)(syntheticSqon));
-  const onLogicalOpChanged = (changedPath, newSqon) =>
-    onSqonChange(setSqonAtPath(changedPath, newSqon)(syntheticSqon));
-  return (
-    <div style={{ display: 'flex', flexDirection: 'row' }}>
-      <div className={`sqonView`}>
-        {isBooleanOp(syntheticSqon) && (
-          <BooleanOp
-            index={0}
-            onFieldOpRemove={onFieldOpRemove}
-            onChange={onLogicalOpChanged}
-            sqon={syntheticSqon}
-          />
-        )}
-      </div>
-    </div>
   );
 };
