+++ modules/components/src/AdvancedSqonBuilder/SqonEntryContent/common.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+export const PillRemoveButton = ({ onClick }) => (
+  <span className={`pillRemoveButton`} onClick={onClick}>
+    ✕
+  </span>
+);
